# Use `vec![x; n]` in BitVec::from_elem