Vector creation optimizes to something like memset properly.

benchmark for from_elem + block traversal:

before:
test bench::bench_from_elem            ... bench:       1,832 ns/iter (+/- 19) = 1117 MB/s
after:
test bench::bench_from_elem            ... bench:         539 ns/iter (+/- 55) = 3799 MB/s